feat: Finish rename from generator to librarian
This includes changing the expected environment variable from
GENERATOR_CLI_REPOSITORY to LIBRARIAN_REPOSITORY.

Additionally, rename the update-repo command to update-apis (to give
space for a later update-image command).
