+++ .github/workflows/librarian.yaml
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-name: Generator
+name: Librarian
 on: [push, pull_request]
 permissions:
   contents: read
